% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_plots.R
\name{colour_on_top}
\alias{colour_on_top}
\title{Reorder paths so selected (colored) paths are on top}
\usage{
colour_on_top(svg_text, ipath_data)
}
\arguments{
\item{svg_text}{text containing svg source code}

\item{ipath_data}{data.frame formatted for sending to ipath. Needs a 'color' column}
}
\value{
pathway map (svg formatted text) with colored paths on top
}
\description{
Reorder paths so selected (colored) paths are on top
}
\author{
Beth Signal
}
\keyword{internal}
