% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{create_ipath_data}
\alias{create_ipath_data}
\title{Format input data to plot with ipath}
\usage{
create_ipath_data(data, id_column, color_column = NA,
  width_column = NA, opacity_column = NA, min_path_width = 3,
  max_path_width = 10, color_type = "discrete", color_cutoff = 0,
  pallete = "RdBu")
}
\arguments{
\item{data}{input data.frame}

\item{id_column}{name of the column containing ids to plot on the ipath map. Must be one of the supported data types and in the correct format (see: https://pathways.embl.de/help.cgi)}

\item{color_column}{name of the column containing values to vary colors in the ipath plot}

\item{width_column}{name of the column containing values to vary width in the ipath plot}

\item{opacity_column}{name of the column containing values to vary opacity in the ipath plot}

\item{min_path_width}{minimum width of paths with a match}

\item{max_path_width}{maximum width of paths with a match. Default width for when no wwidth_column is provided.}

\item{color_type}{type of color palette to map values to (discrete or continous)}

\item{color_cutoff}{value which seperates values into higher or lower for discrete color mapping}

\item{palette}{RColorBrewer pallete name}
}
\value{
data.frame formatted for sending to ipath
}
\description{
Format input data to plot with ipath
}
\author{
Beth Signal
}
