% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_plots.R
\name{find_layer_locs}
\alias{find_layer_locs}
\title{Get indices for the start of each layer in svg text}
\usage{
find_layer_locs(g_splits)
}
\arguments{
\item{g_splits}{vector of svg text split by the delimiter '<g>'}
}
\value{
data.frame of layer numbers and their starting index
}
\description{
Get indices for the start of each layer in svg text
}
\author{
Beth Signal
}
\keyword{internal}
