\name{GSDB_example}
\alias{GSDB_example}
\docType{data}
\title{example gene set database (GSDB)}
\description{
 includes geneset.names, genesets.
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/}
}
\references{
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3106198/}
\format{A list of gene set database}
\usage{data("GSDB_example")}
\examples{
data("GSDB_example")
GSDB_example$geneset.names
}
\keyword{datasets}
