% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iES_cal.R
\name{iES_cal2}
\alias{iES_cal2}
\title{iES calculation Function}
\usage{
iES_cal2(Y, GSDB, BPPARAM = NULL, nPro = 0)
}
\arguments{
\item{Y}{is the expression matrix.}

\item{GSDB}{is the gene set database.}

\item{BPPARAM}{parameters from the BiocParallel.}

\item{nPro}{number of processors (default = 0).}
}
\value{
a matrix with rows corresponding to the pathways and columns corresponding to the patients.
}
\description{
This function calculates the iES matrix which is the core of iPath.
}
\examples{
data(PRAD_data)
data(GSDB_example)
iES_mat = iES_cal2(prad_exprs, GSDB = GSDB_example)
}
\keyword{calculatioin.}
\keyword{iES}
\keyword{statistics}
