% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ipath_pdf.R
\name{make_ipath_ggplot}
\alias{make_ipath_ggplot}
\title{make a ipath ggplot}
\usage{
make_ipath_ggplot(plot_layers)
}
\arguments{
\item{plot_layers}{list of svg layer data produced by svg_to_table()}
}
\description{
make a ipath ggplot
}
\examples{

}
\author{
Beth Signal
}
