% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ipath_pdf.R
\name{make_svg_df}
\alias{make_svg_df}
\title{Convert a svg path specification to data.frame compatible with ggplot
Used for paths with curves (C)}
\usage{
make_svg_df(input, rebase = F)
}
\arguments{
\item{input}{string of svg path}

\item{rebase}{set the minimum path value to 0?}
}
\description{
Convert a svg path specification to data.frame compatible with ggplot
Used for paths with curves (C)
}
\examples{

}
\author{
Beth Signal
}
