% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ipath_pdf.R
\name{make_svg_df_straight}
\alias{make_svg_df_straight}
\title{Convert a svg stright line path specification to data.frame compatible with ggplot
Used for paths without curves (C)}
\usage{
make_svg_df_straight(input)
}
\arguments{
\item{input}{string of svg path}
}
\description{
Convert a svg stright line path specification to data.frame compatible with ggplot
Used for paths without curves (C)
}
\examples{

}
\author{
Beth Signal
}
