% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{map_values_to_colors}
\alias{map_values_to_colors}
\title{Map a vector of values to hex colors}
\usage{
map_values_to_colors(values, color_type = "discrete", color_cutoff = 0,
  pallete = "RdBu")
}
\arguments{
\item{values}{numeric vector of values}

\item{color_type}{type of color palette to map values to (discrete or continous)}

\item{color_cutoff}{value which seperates values into higher or lower for discrete color mapping}

\item{palette}{RColorBrewer pallete name, viridis palette name (if continous), or two colors to create a colorRampPalette from.}
}
\value{
vector of hex colors
}
\description{
Map a vector of values to hex colors
}
\examples{
map_values_to_colors(seq(-4, 5))
map_values_to_colors(seq(-4, 5), color_type = "continous")
map_values_to_colors(seq(-4, 5), color_type = "continous", pallete = "magma")
map_values_to_colors(seq(-4, 5), color_type = "continous", pallete = c("lightblue", "blue"))
}
\author{
Beth Signal
}
\keyword{internal}
