% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{map_values_to_range}
\alias{map_values_to_range}
\title{Map a vector of values to a prespecified range}
\usage{
map_values_to_range(values, min_val = 3, max_val = 10)
}
\arguments{
\item{values}{numeric vector of values}

\item{min_val}{value to map minimum input value to}

\item{max_val}{value to map maximum input value to}
}
\value{
vector of values
}
\description{
Map a vector of values to a prespecified range
}
\examples{
map_values_to_range(seq(0, 10), min_val = 2, max_val=4)

}
\author{
Beth Signal
}
\keyword{internal}
