% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ipath_pdf.R
\name{print_ipath_pdf}
\alias{print_ipath_pdf}
\title{Write an iPath3 metabolic pathways map to pdf}
\usage{
print_ipath_pdf(ipath_data, file_name = NULL, default_width = 2,
  default_color = "#bdbdbd", default_radius = 5,
  textbox_textcol = NULL, textbox_col = NULL)
}
\arguments{
\item{ipath_data}{data.frame formatted for sending to ipath}

\item{file_name}{pdf file name to export plot.}

\item{default_width}{Default path/edge width (px) for the non-selected parts of the map.}

\item{default_color}{Hex color value. Non-selected parts of the map will use this color by default.}

\item{default_radius}{Compound radius (px) for the non-selected parts of the map.}

\item{textbox_textcol}{hex value for text (in box) color}

\item{textbox_col}{hex value for box color}
}
\description{
Write an iPath3 metabolic pathways map to pdf
}
\author{
Beth Signal
}
