% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_plots.R
\name{recolor_textbox}
\alias{recolor_textbox}
\title{Change color of boxes surrounding major pathway names}
\usage{
recolor_textbox(pw_map, box_color = "#969696", text_color = "#FFFFFF")
}
\arguments{
\item{box_color}{hex value for box color}

\item{text_color}{hex value for text (in box) color}

\item{pathway_map}{text containing svg source code}
}
\value{
pathway map with altered box colors
}
\description{
Change color of boxes surrounding major pathway names
}
\author{
Beth Signal
}
\keyword{internal}
