% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{rem_data}
\alias{rem_data}
\title{remove genes with 0 sd}
\usage{
rem_data(Y)
}
\arguments{
\item{Y}{is the expression matrix.}
}
\value{
a processed matrix
}
\description{
This function helps remove non-informative genes.
}
