% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ipath_pdf.R
\name{svg_to_table}
\alias{svg_to_table}
\title{Convert svg data to tables for ggplot}
\usage{
svg_to_table(svg_xml)
}
\arguments{
\item{svg_xml}{data.frame containing svg data}
}
\description{
Convert svg data to tables for ggplot
}
\author{
Beth Signal
}
