% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{to_parameters}
\alias{to_parameters}
\title{Format ipath data.frame as parameters}
\usage{
to_parameters(ipath_data, default_color = "#bdbdbd", default_width = 3,
  default_radius = 7, default_opacity = 1, keep_colors = FALSE,
  whole_modules = FALSE, whole_pathways = FALSE,
  query_reactions = FALSE, tax_filter = "", export_dpi = 1200)
}
\arguments{
\item{ipath_data}{data.frame formatted for sending to ipath}

\item{default_color}{Hex color value. Non-selected parts of the map will use this color by default.}

\item{default_width}{Default path/edge width (px) for the non-selected parts of the map.}

\item{default_radius}{Compound radius (px) for the non-selected parts of the map.}

\item{default_opacity}{Opacity setting for the non-selected parts of the map. Zero represents full transparency, while one is fully opaque.}

\item{keep_colors}{Keep original colors? When TRUE, non-selected parts of the map will not change color to the default one.}

\item{whole_modules}{Select whole modules? When TRUE, any KEGG module with at least one matching edge or compound will be highlighted.}

\item{whole_pathways}{Select whole pathways? If TRUE, any pathway with at least one matching edge or compound will be highlighted.}

\item{query_reactions}{Query reaction compounds. If TRUE, compound presence within each edges reactions will also be checked}

\item{tax_filter}{string containing either NCBI tax ID or KEGG 3 letter species code(s). Only pathways present in selected species will be included in the map.}

\item{export_dpi}{numeric value for DPI of SVG output.}
}
\value{
list of parameters for ipath
}
\description{
Format ipath data.frame as parameters
}
\author{
Beth Signal
}
\keyword{internal}
